+++ lib/menus.dart
@@ -1,6 +1,7 @@
 /// Provides exports relating to menus.
 library menus;
 
+export 'src/menu/asset_reference_menu.dart';
 export 'src/menu/file_picker_menu.dart';
 export 'src/menu/menu.dart';
 export 'src/menu/menu_item.dart';